// SMSFCart/SMSCartWeb/src/app/models/VendescModel.ts
export class Vendesc {
  itm_size:string;
  material_type: string ;
  itm_wtid:number ;
  worktype:string ;
  itm_id: number;
 
  itm_descp: string;
  care: string;
  attached_sleeves: string;
  sleeve_Length: number;
  sleeves_material: string;
  price: number;
  color: string;
  product_code: string;
  shipping_charges: number;
  offer: number;
  bid: number;

  //items_sid: number;
  item_mid: number;
  itm_sz_id: number;
  Bname: string;
  itm_length: number;
  quantity: number;
  features: string;
  occasion: string;
  itm_style_type: string;
  neck_type: string;

  toe_type: string;
  heel_height: string;
  heel_type: string;
  model_no: string;
  it_name: string;
  waranty: number;
}